do not use rich in notebooks
