+++ nilearn/_utils/helpers.py
@@ -230,3 +230,34 @@ def is_gil_enabled() -> bool:
 
 def is_sphinx_build() -> bool:
     return any(module.startswith("sphinx.") for module in sys.modules)
+
+
+def is_notebook() -> bool:
+    """Detect if we are running in a notebook.
+
+    Adapted from https://stackoverflow.com/questions/15411967/how-can-i-check-if-code-is-executed-in-the-ipython-notebook
+    """
+    try:
+        shell = get_ipython().__class__.__name__  # type: ignore[name-defined]
+    except NameError:
+        shell = False
+
+    try:
+        import marimo as mo
+
+        is_marimo = mo.running_in_notebook()
+    except ImportError:
+        is_marimo = False
+
+    if shell:
+        if shell == "ZMQInteractiveShell":
+            return True  # Jupyter notebook or qtconsole
+        elif shell == "TerminalInteractiveShell":
+            return False  # Terminal running IPython
+        else:
+            return False  # Other type (?)
+
+    if is_marimo:
+        return is_marimo
+
+    return False
