+++ nilearn/glm/_reporting_utils.py
@@ -10,7 +10,7 @@ from sklearn.utils import Bunch
 
 from nilearn import DEFAULT_DIVERGING_CMAP
 from nilearn._utils.docs import fill_doc
-from nilearn._utils.helpers import is_matplotlib_installed
+from nilearn._utils.helpers import is_matplotlib_installed, is_notebook
 from nilearn._utils.logger import find_stack_level
 from nilearn._utils.niimg import safe_get_data
 from nilearn._utils.param_validation import (
@@ -24,10 +24,7 @@ from nilearn.reporting.get_clusters_table import (
     clustering_params_to_dataframe,
     get_clusters_table,
 )
-from nilearn.reporting.html_report import (
-    HTMLReport,
-    is_notebook,
-)
+from nilearn.reporting.html_report import HTMLReport
 from nilearn.reporting.mixin import ReportMixin
 from nilearn.reporting.utils import figure_to_png_base64
 from nilearn.surface.surface import SurfaceImage
